Web client can now handle logging in from another browser; terminal chats shut down if session is cancelled.
If the same user logs into a different browser, the existing avatar shuts down
and signals its client to navigate to /logout.  This isn't always possible
(e.g. if the browser tab was closed).  In that case, try to send the signal and
move on.

/logout ends the session and redirects to /login .

For terminal users, if you have a chat dialog open and the session is
cancelled, the chat won't work anymore (you need to at least be invited to a
session).  In this case, the chat window is immediately closed.

This is superior to the previous behavior, which left the dialog drawn on the
terminal screen but all the keys typed were routed to the application input
handler.