Handle failure to write to the event source.
This can happen if you walk away from the game and the connection from
the browser to the server times out.  Maybe a network error.

The result is that user needs to refresh their browser, but the game
shouldn't be stuck until that happens.  Instead, the event source
should be reset and events buffered.